+++ usps_track/__init__.py
@@ -1,4 +1,4 @@
 """usps-track module."""
 from __future__ import annotations
 
-__version__ = '0.0.4'
+__version__ = '0.0.5'