+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.1/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [Unreleased]
+
+## [0.0.5] - 2026-04-27
 
 ### Changed
 
@@ -43,7 +45,8 @@ Maintenance release.
 
 First version.
 
-[unreleased]: https://github.com/Tatsh/usps-track/compare/v0.0.4...HEAD
+[unreleased]: https://github.com/Tatsh/usps-track/compare/v0.0.5...HEAD
+[0.0.5]: https://github.com/Tatsh/usps-track/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/Tatsh/usps-track/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/Tatsh/usps-track/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/Tatsh/usps-track/compare/v0.0.1...v0.0.2